Rename package`duke.task` to `duke.tasks`
