+++ src/main/java/duke/Storage.java
@@ -9,10 +9,10 @@ import java.nio.file.Paths;
 import java.util.Scanner;
 
 import duke.exceptions.CorruptedSaveException;
-import duke.task.Deadline;
-import duke.task.Event;
-import duke.task.Task;
-import duke.task.ToDo;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.ToDo;
 
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
