+++ src/main/java/duke/tasks/Event.java
@@ -1,4 +1,4 @@
-package duke.task;
+package duke.tasks;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
@@ -8,33 +8,34 @@ import duke.exceptions.CorruptedSaveException;
 import duke.exceptions.DukeException;
 
 /**
- * Represents a Deadline which is a kind of Task.
+ * Represents an Event which is a kind of Task.
  * Encapsulates an additional LocalDate and LocalTime attribute
- * which represents the deadline of the Task.
+ * which represents when the Event occur.
  */
-public class Deadline extends Task {
+public class Event extends Task {
     private static final DateTimeFormatter TIME_FORMATTER = DateTimeFormatter.ofPattern("HHmm");
     private static final DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+    private LocalDate startDate;
     private LocalTime time;
-    private LocalDate deadline;
 
     /**
-     * Creates a new Deadline that is marked as not done.
+     * Creates a new Event that is marked as not done by default.
      * 
-     * <p>Accepts a deadlineDate string with pattern "yyyy-MM-dd" or "yyyy-MM-dd HHmm".</p>
+     * <p>Accepts a startDateString with pattern "yyyy-MM-dd" or "yyyy-MM-dd HHmm".</p>
      * 
      * @param task The description of the task.
-     * @param deadlineDate The string representation of the deadline.
-     * @throws DukeException If deadlineDate does not have the expected pattern.
+     * @param startDateString The string representation of when the event occurs.
+     * @throws DukeException If startDateString does not have the expected pattern.
      */
-    public Deadline(String task, String deadlineDate) throws DukeException {
+    public Event (String task, String startDateString) throws DukeException {
         super(task);
-        String[] datetime = deadlineDate.split(" ");
-
+        String[] datetime = startDateString.split(" ");
+        
         try {
-            deadline = LocalDate.parse(datetime[0], DATE_FORMATTER);
+            startDate = LocalDate.parse(datetime[0], DATE_FORMATTER);
             time = datetime.length == 1 ? null : LocalTime.parse(datetime[1], TIME_FORMATTER);
         } catch (DateTimeParseException e) {
+            System.out.println(e);
             throw new DukeException("Invalid date/time format!" 
                     + " Expected date and/or time in the following formats: \n"
                     + "yyyy-mm-dd | Example: 2022-06-26\n"
@@ -43,80 +44,78 @@ public class Deadline extends Task {
     }
 
     /**
-     * Creates a new Deadline that is marked as not done.
+     * Creates a new Event that is marked as not done.
+     * 
      * 
      * @param task The description of the task.
-     * @param deadline Represents the date by which the task is due. 
-     * @param time Represents the time (if any) by which the task is due. 
+     * @param startDate Represents the date at which the event occurs. 
+     * @param time Represents the time (if any) by which the event occurs. 
      */
-    public Deadline(String task, LocalDate deadline, LocalTime time) {
+    public Event(String task, LocalDate startDate, LocalTime time) {
         super(task);
-        this.deadline = deadline;
+        this.startDate = startDate;
         this.time = time;
     }
 
     /**
-     * Constructor to create a Deadline instance.
+     * Constructor to create an Event instance.
      * 
      * <p>This constructor accepts an additional isDone boolean to initialize 
      * a task that has been marked/unmarked as done. </p>
      * 
-     * <p>Accepts a deadlineDate string with pattern "yyyy-MM-dd HHmm" or "yyyy-MM-dd".</p>
-     * 
+     * <p>Accepts a startDate string with pattern "yyyy-MM-dd HHmm" or "yyyy-MM-dd".</p>
      * @param task The description of the task.
      * @param isDone Marks the task as done if true.
-     * @param deadlineDate The string representation of the deadline.
-     * @throws DukeException If deadlineDate does not have the expected pattern.
+     * @param startDate The string representation of when the event occurs.
+     * @throws DukeException If startDate does not have the expected pattern.
      */
-    public Deadline (String task, boolean isDone, String deadlineDate) throws DukeException {
-        this(task, deadlineDate);
+    public Event (String task, boolean isDone, String startDate) throws DukeException {
+        this(task, startDate);
         this.isDone = isDone;
     }
 
     /**
-     * Formats a Deadline instance to be stored in an external file.
+     * Formats an Event instance to be stored in an external file.
      */
     @Override
     public String toFileFormat() {
         Integer i = this.isDone ? 1 : 0;
-        String deadlineString = deadline.format(DATE_FORMATTER);
+        String startDateString = startDate.format(DATE_FORMATTER);
         String timeString = time == null ? "" : " " + time.format(TIME_FORMATTER);
-        return String.format("D | %d | %s | %s%s\n", 
-                i, this.task, deadlineString, timeString);
+        return String.format("E | %d | %s | %s%s\n", 
+                i, this.task, startDateString, timeString);
     }
 
     /**
      * Converts a string from file format to Task.
      * 
-     * <p> The inverse of toFileFormat().</p>
+     * <p> The inverse of `toFileFormat()`.</p>
      * 
      * @param fileString The string to convert.
      * @throws CorruptedSaveException if unable to parse the string correctly
      */
-    public static Deadline fromFileFormat(String fileString) throws CorruptedSaveException {
+    public static Event fromFileFormat(String fileString) throws CorruptedSaveException {
         String[] packetSections = fileString.split(" \\| ");
         try {
             boolean isDone = Integer.parseInt(packetSections[1]) == 1;
             String taskName = packetSections[2];
-            String deadlineString = packetSections[3];
-            return new Deadline(taskName, isDone, deadlineString);
+            String startDateString = packetSections[3];
+            return new Event(taskName, isDone, startDateString);
         } catch (Exception e) {
             throw new CorruptedSaveException();
         }
-
     }
 
     /**
-     * Returns the string representation of a Deadline.
+     * Returns the string representation of an Event.
      */
     @Override
     public String toString() {
         DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("MMM dd yyyy");
         DateTimeFormatter timeFormatter = DateTimeFormatter.ofPattern("HH:mm");
 
-        String date = this.deadline.format(dateFormatter);
+        String date = this.startDate.format(dateFormatter);
         String time = this.time == null ? "" : " " + this.time.format(timeFormatter);
-        return String.format("[D]%s %s (by: %s%s)", this.statusString(), this.task, date, time);
+        return String.format("[E]%s %s (at: %s%s)", this.statusString(), this.task, date, time);
     }
-
 }
