+++ src/main/java/duke/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package duke.task;
+package duke.tasks;
 
 import duke.exceptions.CorruptedSaveException;
