+++ src/test/java/duke/TaskStub.java
@@ -1,6 +1,6 @@
 package duke;
 
-import duke.task.Task;
+import duke.tasks.Task;
 
 public class TaskStub extends Task {
